Fix the axis position
If not explicitly set, axis position of "xaxis" is displayed against
"yaxis". As the axis order is not strictly to have left, right, bottom
or top in the beginning, referencing the main axis, this leads to
wrongly set axes.

Default to plotting the left axis at 0% and the right axis at 100%

This does not respect axes that are drawn at specific percentages.