Allow Saving Of Output on Arbitrary Paths
Previously, all images were saved in the "home" path of the Experiment.
No subfolders were allowed and the save to absolute folders could not
work.

This extends the file path handling and creates an Igor Path as a
symbolic reference on the file system. This is especially useful if you
you want to save to a subfolder